+++ App.js
@@ -28,7 +28,6 @@ import SearchResults from './screens/search/SearchResults';
 import RoomAgenda from './screens/room/RoomAgenda';
 import ChatScreen from './screens/chat/Chat';
 import ComponentsExample from './screens/ComponentsExample';
-import ChatScreenTest from './screens/chat/Chat_test';
 import ContactScreen from './screens/chat/Contacts';
 
 const Stack = createNativeStackNavigator();
@@ -98,7 +97,7 @@ export default function App() {
         <Stack.Screen name='RoomAgenda' component={RoomAgenda} />
         <Stack.Screen name='ChatScreen' component={ChatScreen} />
         <Stack.Screen name='Contacts' component={ContactScreen} />
-        <Stack.Screen name='ChatScreenTest' component={ChatScreenTest} />
+        {/* <Stack.Screen name='ChatScreenTest' component={ChatScreenTest} /> */}
       </Stack.Navigator>
     </NavigationContainer>
   );
