+++ screens/chat/Contacts.js
@@ -36,8 +36,8 @@ export default function ContactScreen({ navigation }) {
         const userID = userIDvalue.replace(/['"]+/g, '');
 
         try {
-            const endpoint = SERVER_URL + '/users/all';
-
+            //const endpoint = SERVER_URL + '/users/all';
+            const endpoint = SERVER_URL + `/chat/dms/${userID}`;
             const response = await fetch(endpoint, {
                 method: 'GET',
                 headers: requestHeaders
@@ -45,15 +45,13 @@ export default function ContactScreen({ navigation }) {
 
             const users = await response.json();
             //console.log(users);
-
-            const otherUsers = []
-            for( const user of users){
-                if(String(user._id) != userID){
-                    otherUsers.push(user);
-                }
-            }
-
-            setUsers(otherUsers);
+            // const otherUsers = []
+            // for(const user of users){
+            //     if(String(user._id) != userID){
+            //         otherUsers.push(user);
+            //     }
+            // }
+            setUsers(users);
         } catch (error) {
             console.error(error);
             alert('Something went wrong');
@@ -67,16 +65,15 @@ export default function ContactScreen({ navigation }) {
     const renderUsers = ({ item }) => (
         <TouchableOpacity
             style={styles.mainProfileButton}
-            onPress={() => navigation.navigate('ChatScreenTest', { _id: item._id, username: item.credentials.username })}>
-            
+            onPress={() => navigation.navigate('ChatScreen', { _id: item._id, username: item.credentials.username })}>
             <View style={styles.pfpWrapper}>
                 <ProfileIcon
                     image={(item.profile_pic ? {uri: item.profile_pic} : require('../../assets/images/Avatar.png'))}
                 />
             </View>
             <View style={{flexDirection: 'column'}}>
-            <Text style={[styles.buttonText, textStyle.h2]}>{item.name.first_name} {item.name.last_name}</Text>
-            <Text style={[styles.subtitle, textStyle.h3]}>{item.credentials.username}</Text>
+                <Text style={[styles.buttonText, textStyle.h2]}>{item.name.first_name} {item.name.last_name}</Text>
+                <Text style={[styles.subtitle, textStyle.h3]}>{item.credentials.username}</Text>
             </View>
         </TouchableOpacity>
     );
@@ -85,7 +82,7 @@ export default function ContactScreen({ navigation }) {
     return(
         <SafeAreaView style={styles.container}>
             <View style={styles.header}>
-                <Text style={textStyle.h1}>Users</Text>
+                <Text style={textStyle.h1}>Your messages</Text>
             </View>
             <FlatList
                 data={users}
